back-end/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
PyJWT
Werkzeug
Flask-Migrate

